+++ z0bug_odoo/travis/getaddons.py
@@ -20,7 +20,7 @@ else:
     except ImportError:
         from git_run import GitRun
 
-__version__ = '1.0.12.1'
+__version__ = '1.0.13'
 
 MANIFEST_FILES = [
     '__manifest__.py',
@@ -28,7 +28,6 @@ MANIFEST_FILES = [
     '__openerp__.py',
     '__terp__.py',
 ]
-
 
 
 def is_module(path):