+++ zerobug/tests/zerobug_test_05.py
@@ -10,7 +10,7 @@ import os
 import sys
 from zerobug import z0test
 
-__version__ = "1.0.8.2"
+__version__ = "1.0.9"
 
 MODULE_ID = 'zerobug'
 TEST_FAILED = 1