+++ clodoo/egg-info/history.rst
@@ -1,5 +1,5 @@
-1.0.1.2 (2022-03-29)
-~~~~~~~~~~~~~~~~~~~~
+1.0.2 (2022-03-29)
+~~~~~~~~~~~~~~~~~~
 
 * [FIX] transodoo.xlsx: account.invoice-move
 * [FIX] transodoo.py: account.invoice-move