+++ egg-info/history.rst
@@ -1,5 +1,32 @@
-zerobug: 1.0.8.2 (2022-03-25)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+odoo_score: 1.0.6 (2022-04-09)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] run_odoo_debug: test with coverage
+
+
+wok_code: 1.0.9 (2022-04-04)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] pull_repo.sh: improvements
+* [IMP] Remove old unused code
+* [IMP] please: new actions lint and test
+
+
+z0bug_odoo: 1.0.13 (2022-03-29)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] Invoice data
+
+
+clodoo: 1.0.2 (2022-03-29)
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] transodoo.xlsx: account.invoice-move
+* [FIX] transodoo.py: account.invoice-move
+
+
+zerobug: 1.0.9 (2022-03-25)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 * [FIX] Sometime fails test result
 * [IMP] Lint checks
@@ -20,9 +47,8 @@ clodoo: 1.0.1.1 (2022-03-25)
 * [IMP] more tests
 
 
-
-python_plus: 1.0.8.1 (2022-03-24)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+python_plus: 1.0.9 (2022-03-24)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 * [IMP] Lint + tests
 
@@ -40,8 +66,8 @@ zerobug: 1.0.8.1 (2022-03-19)
 * [IMP] black formatting
 
 
-z0lib: 1.0.7.1 (2022-03-19)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~
+z0lib: 1.0.8 (2022-03-19)
+~~~~~~~~~~~~~~~~~~~~~~~~~
 
 * [IMP] More tests